+++ django_migration_linter/sql_analyser/__init__.py
@@ -3,4 +3,4 @@ from .mysql import MySqlAnalyser  # noqa
 from .postgresql import PostgresqlAnalyser  # noqa
 from .sqlite import SqliteAnalyser  # noqa
 
-from .analyser import analyse_sql_statements  # noqa isort:skip
+from .analyser import analyse_sql_statements, get_sql_analyser_class  # noqa isort:skip