+++ django_migration_linter/sql_analyser/analyser.py
@@ -1,7 +1,6 @@
 import logging
 
 from django_migration_linter.sql_analyser import (
-    BaseAnalyser,
     MySqlAnalyser,
     PostgresqlAnalyser,
     SqliteAnalyser,
@@ -9,8 +8,28 @@ from django_migration_linter.sql_analyser import (
 
 logger = logging.getLogger("django_migration_linter")
 
+ANALYSER_STRING_MAPPING = {
+    "sqlite": SqliteAnalyser,
+    "mysql": MySqlAnalyser,
+    "postgresql": PostgresqlAnalyser,
+}
 
-def get_sql_analyser(database_vendor, exclude_migration_tests=None):
+
+def get_sql_analyser_class(database_vendor, analyser_string=None):
+    if analyser_string:
+        return get_sql_analyser_from_string(analyser_string)
+    return get_sql_analyser_class_from_db_vendor(database_vendor)
+
+
+def get_sql_analyser_from_string(analyser_string):
+    if analyser_string not in ANALYSER_STRING_MAPPING:
+        raise ValueError(
+            "Unknown SQL analyser. Known values: {}".format(analyser_string.keys())
+        )
+    return ANALYSER_STRING_MAPPING[analyser_string]
+
+
+def get_sql_analyser_class_from_db_vendor(database_vendor):
     if "mysql" in database_vendor:
         sql_analyser_class = MySqlAnalyser
     elif "postgre" in database_vendor:
@@ -18,15 +37,15 @@ def get_sql_analyser(database_vendor, exclude_migration_tests=None):
     elif "sqlite" in database_vendor:
         sql_analyser_class = SqliteAnalyser
     else:
-        sql_analyser_class = BaseAnalyser
+        raise ValueError("Unsupported database vendor. Try specifying an SQL analyser.")
 
     logger.debug("Chosen SQL analyser class: %s", sql_analyser_class)
-    return sql_analyser_class(exclude_migration_tests)
+    return sql_analyser_class
 
 
 def analyse_sql_statements(
-    sql_statements, database_vendor, exclude_migration_tests=None
+    sql_analyser_class, sql_statements, exclude_migration_tests=None
 ):
-    sql_analyser = get_sql_analyser(database_vendor, exclude_migration_tests)
+    sql_analyser = sql_analyser_class(exclude_migration_tests)
     sql_analyser.analyse(sql_statements)
     return sql_analyser.errors, sql_analyser.ignored, sql_analyser.warnings